Initial changes for moving authplugin into its own separate module
