+++ controller/src/test/java/io/pravega/controller/auth/PasswordEncryptorTool.java
@@ -9,7 +9,7 @@
  */
 package io.pravega.controller.auth;
 
-import io.pravega.controller.server.security.auth.StrongPasswordProcessor;
+import io.pravega.shared.security.StrongPasswordProcessor;
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
