+++ controller/src/test/java/io/pravega/controller/server/rpc/auth/RESTAuthHelperTest.java
@@ -16,7 +16,7 @@ import static io.pravega.auth.AuthHandler.Permissions.*;
 import io.pravega.auth.AuthException;
 import io.pravega.controller.mocks.FakeAuthHandler;
 import io.pravega.controller.server.security.auth.RESTAuthHelper;
-import io.pravega.controller.server.security.auth.UserPrincipal;
+import io.pravega.shared.security.UserPrincipal;
 import io.pravega.controller.server.security.auth.handler.AuthHandlerManager;
 import org.junit.Before;
 import org.junit.Test;
