+++ settings.gradle
@@ -14,6 +14,7 @@ include 'common',
         'common_server',
         'shared:cluster',
         'shared:authplugin',
+        'shared:basic-authplugin',
         'shared:security',
         'shared:metrics',
         'shared:protocol',
@@ -32,3 +33,6 @@ include 'common',
         'test:integration',
         'test:system',
         'bindings'
+include 'password-authplugin'
+include 'basic-authplugin'
+
