+++ shared/basic-authplugin/src/main/java/io/pravega/authplugin/basic/AccessControlList.java
@@ -7,7 +7,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.controller.server.security.auth.handler.impl;
+package io.pravega.authplugin.basic;
 
 import lombok.AccessLevel;
 import lombok.Getter;
