+++ shared/basic-authplugin/src/main/resources/META-INF/services/io.pravega.auth.AuthHandler
@@ -7,4 +7,4 @@
 #
 #     http://www.apache.org/licenses/LICENSE-2.0
 #
-io.pravega.controller.server.security.auth.handler.impl.PasswordAuthHandler
+io.pravega.authplugin.basic.PasswordAuthHandler
